Tests/LibWeb/Text/expected/resolve-css-overflow-effective-value.txt

overflow-x: auto
overflow-y: hidden
